Brick breaking particle effect
This adds a particle effect for when you break bricks. Also, GetBlockX
was missing an RTS which was causing problems. Touching springs also
causes a jump to start, so you can be launched into bricks and break
them. There was a pla in the main loop that was not needed and was
probably messing up the stack.